requirements.txt
pandas
numpy
matplotlib
sklearn
python-dateutil
xlrd
openpyxl
Unidecode
